% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{decodeAndSave}
\alias{decodeAndSave}
\title{Decode encoded base64 then save it as png}
\usage{
decodeAndSave(encoded_, to_ = "mypng.png")
}
\value{
a png picture
}
\description{
Decode encoded base64 then save it as png
}
\keyword{internal}
