% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_hosting.R
\name{edit_environ_gitee}
\alias{edit_environ_gitee}
\title{Add necessary configurations in ~/.Renviron for gitee.}
\usage{
edit_environ_gitee()
}
\value{
Return null but open `~/.Renviron` then insert configurations.
}
\description{
Add necessary configurations in ~/.Renviron for gitee.
}
