% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_hosting.R
\name{edit_environ_github}
\alias{edit_environ_github}
\title{Add necessary configurations in ~/.Renviron for github.}
\usage{
edit_environ_github()
}
\value{
Return null but open `~/.Renviron` then insert configurations.
}
\description{
Add necessary configurations in ~/.Renviron for github.
}
