% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_hosting.R
\name{edit_environ_githubAction}
\alias{edit_environ_githubAction}
\title{Add necessary configurations in ~/.Renviron for github action service.}
\usage{
edit_environ_githubAction()
}
\value{
Return null but open `~/.Renviron` then insert configurations.
}
\description{
Add necessary configurations in ~/.Renviron for github action service.
}
