% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyapp.R
\name{insert_pic}
\alias{insert_pic}
\title{Main UI function}
\usage{
insert_pic()
}
\value{
Inserts selected markdown/knitr link at current location and saves it in designated path.
}
\description{
Main UI function
}
\examples{
\dontrun{
 kaca()
}

}
\keyword{internal}
