% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{kaca}
\alias{kaca}
\title{Invoke RStudio add-in to save and insert pictures in clipboard}
\usage{
kaca()
}
\value{
Inserts selected markdown/knitr link at current location and saves it in designated path.
}
\description{
Invoke RStudio add-in to save and insert pictures in clipboard
}
\details{

}
\examples{
\dontrun{
 kaca()
}

}
