% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_hosting.R
\name{remoteFilepath}
\alias{remoteFilepath}
\title{remoteFilepath}
\usage{
remoteFilepath(where_ = tolower(Sys.getenv("HOST_SERVICE")), filename_)
}
\value{
path to file in repos
}
\description{
remoteFilepath
}
\keyword{internal}
