% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_hosting.R
\name{upload2gitee}
\alias{upload2gitee}
\title{upload2gitee}
\usage{
upload2gitee(filename_, content_)
}
\value{
Return null but upload file to repos(gitee)
}
\description{
upload2gitee
}
\keyword{internal}
