% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_hosting.R
\name{upload2github}
\alias{upload2github}
\title{upload2github}
\usage{
upload2github(filename_, content_)
}
\value{
Return null but upload file to repos(github)
}
\description{
upload2github
}
\keyword{internal}
