% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_hosting.R
\name{upload2where}
\alias{upload2where}
\title{upload2where}
\usage{
upload2where(
  where_ = tolower(Sys.getenv("HOST_SERVICE")),
  filename_,
  base64_ = NULL,
  src_ = NULL
)
}
\value{
Inserts selected markdown/knitr link at current location and saves it in designated path.
}
\description{
upload2where
}
\keyword{internal}
